chore: add pyright for type checking (#35)
## How does this PR impact the user?

This PR will speed up development and reduce bugs thanks to the type
checking.

## Description

- [x] setup pyright for type checking
- [x] add extra types or type checks here and there
- [x] fix a bug in the `_trim_google_meet_links_from_description` where
it could've returned `None` instead of the calendar event description

## Limitations

The typing is a bit relaxed around the generic types to allow us to use
telegram bot types without providing all of the generic arguments. We
should probably circle back to this and fix the types.

## Checklist

- [x] my PR is focused and contains one wholistic change
- [ ] I have added screenshots or screen recordings to show the changes